Slim down dependencies required by API modules. Extract web service implementations to microservices that need them.
