+++ course-manager/build.gradle.kts
@@ -1,4 +1,7 @@
 plugins {
+    id("io.spring.dependency-management")
+    id("org.springframework.boot")
+
     kotlin("jvm")
     kotlin("plugin.spring")
     kotlin("plugin.jpa")
