+++ department-manager/src/main/kotlin/com/pearlbailey/departmentmanager/downstream/DefaultTeacherWebService.kt
@@ -1,10 +1,11 @@
-package com.pearlbailey.teachermanager.api
+package com.pearlbailey.departmentmanager.downstream
 
 import com.pearlbailey.commontools.web.AbstractWebService
 import com.pearlbailey.teachermanager.api.TeacherConstants.TEACHERS_RESOURCE_PATH
 import com.pearlbailey.teachermanager.api.model.web.CreateTeacherDto
 import com.pearlbailey.teachermanager.api.model.web.GetTeacherResponseDto
 import com.pearlbailey.teachermanager.api.model.web.PatchTeacherDto
+import com.pearlbailey.teachermanager.api.service.TeacherWebService
 import org.springframework.web.client.RestTemplate
 
 class DefaultTeacherWebService(teacherManagerUrl: String, restTemplate: RestTemplate) :
