+++ department-manager/src/test/kotlin/com/pearlbailey/departmentmanager/DefaultDepartmentServiceTest.kt
@@ -4,8 +4,8 @@ import com.pearlbailey.departmentmanager.api.DepartmentFactory
 import com.pearlbailey.departmentmanager.api.model.CreateDepartmentDto
 import com.pearlbailey.departmentmanager.api.model.Department
 import com.pearlbailey.teachermanager.api.TeacherFactory
-import com.pearlbailey.teachermanager.api.TeacherWebService
 import com.pearlbailey.teachermanager.api.model.web.TeacherNotFoundException
+import com.pearlbailey.teachermanager.api.service.TeacherWebService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
