+++ student-manager-api/build.gradle.kts
@@ -1,3 +1,5 @@
+import org.springframework.boot.gradle.tasks.bundling.BootJar
+
 plugins {
     kotlin("jvm")
 }
@@ -6,13 +8,15 @@ java.sourceCompatibility = JavaVersion.VERSION_17
 
 dependencies {
     implementation(project(":common-tools"))
-    implementation("javax.persistence:javax.persistence-api:2.2")
 
-    implementation("org.springdoc:springdoc-openapi-ui:1.6.4")
-    implementation("org.springdoc:springdoc-openapi-kotlin:1.6.4")
+    implementation("com.fasterxml.jackson.core:jackson-databind:2.13.3")
+    implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.3")
+    implementation("javax.persistence:javax.persistence-api:2.2")
+    implementation("io.swagger.core.v3:swagger-annotations:2.1.12")
+    implementation("javax.validation:validation-api:2.0.1.Final")
 }
 
-tasks.getByName<org.springframework.boot.gradle.tasks.bundling.BootJar>("bootJar") {
+tasks.getByName<BootJar>("bootJar") {
     enabled = false
 }
