+++ teacher-manager-api/src/main/kotlin/com/pearlbailey/teachermanager/api/model/web/TeacherNotFoundException.kt
@@ -1,6 +1,6 @@
 package com.pearlbailey.teachermanager.api.model.web
 
-import org.webjars.NotFoundException
+import com.pearlbailey.commontools.exception.NotFoundException
 
 class TeacherNotFoundException(teacherId: Int? = null, message: String? = null) :
     NotFoundException(message ?: "Teacher with id $teacherId not found.")
